Moves source files to common structure for packaging
