+++ mqtt_monitor.py
@@ -22,15 +22,15 @@ from typing import Any
 
 from aiohttp import ClientSession
 from aiohttp.client_exceptions import ClientError
-from api.api import AnkerSolixApi  # pylint: disable=no-name-in-module
-from api.apitypes import Color  # pylint: disable=no-name-in-module
-from api.errors import AnkerSolixError  # pylint: disable=no-name-in-module
-from api.mqtt import (  # pylint: disable=no-name-in-module
+from anker_solix_api.api import AnkerSolixApi  # pylint: disable=no-name-in-module
+from anker_solix_api.apitypes import Color  # pylint: disable=no-name-in-module
+from anker_solix_api.errors import AnkerSolixError  # pylint: disable=no-name-in-module
+from anker_solix_api.mqtt import (  # pylint: disable=no-name-in-module
     AnkerSolixMqttSession,
     MessageCallback,
 )
-from api.mqtt_factory import SolixMqttDeviceFactory  # pylint: disable=no-name-in-module
-from api.mqtttypes import (  # pylint: disable=no-name-in-module
+from anker_solix_api.mqtt_factory import SolixMqttDeviceFactory  # pylint: disable=no-name-in-module
+from anker_solix_api.mqtttypes import (  # pylint: disable=no-name-in-module
     DeviceHexData,
     DeviceJsonData,
 )
