+++ src/anker_solix_api/__init__.py
@@ -1,0 +1,4 @@
+"""Init for api."""
+
+# import for backward compatibility
+from anker_solix_api import *
