+++ src/anker_solix_api/apitypes.py
