+++ src/anker_solix_api/errors.py
