+++ src/anker_solix_api/mqtt_generator.py
