+++ src/anker_solix_api/mqttcmdmap.py
