+++ src/anker_solix_api/mqtttypes.py
