+++ src/anker_solix_api/schedule.py
